move blockchain util functions
